requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.1
logging==0.4.9.6 